+++ Runtime/Gesture/WindowsMRGestures.cs
@@ -49,17 +49,26 @@ namespace UnityEngine.XR.WindowsMR
         public bool enableNavigationGesture = false;
         public bool enableManipulationGesture = false;
 
-        void Start()
+        WindowsMRGestureSubsystem GetGestureSubsystemIfNeeded()
         {
-            List<XRGestureSubsystem> gestureSubsystems = new List<XRGestureSubsystem>();
-            SubsystemManager.GetInstances<XRGestureSubsystem>(gestureSubsystems);
-            foreach (var subsystem in gestureSubsystems)
+            if (gestureSubsystem == null)
             {
-                if (subsystem != null && subsystem is WindowsMRGestureSubsystem)
-                    gestureSubsystem = subsystem as WindowsMRGestureSubsystem;
+                List<XRGestureSubsystem> gestureSubsystems = new List<XRGestureSubsystem>();
+                SubsystemManager.GetInstances<XRGestureSubsystem>(gestureSubsystems);
+                foreach (var subsystem in gestureSubsystems)
+                {
+                    if (subsystem != null && subsystem is WindowsMRGestureSubsystem)
+                        gestureSubsystem = subsystem as WindowsMRGestureSubsystem;
+                }
+
+                OnValidate();
             }
 
-            OnValidate();
+            return gestureSubsystem;
+        }
+
+        void Start()
+        {
         }
 
         void OnValidate()
@@ -76,38 +85,40 @@ namespace UnityEngine.XR.WindowsMR
 
         void Update()
         {
-            if (gestureSubsystem == null || !gestureSubsystem.running)
+            var gesturesSub = GetGestureSubsystemIfNeeded();
+
+            if (gesturesSub == null || !gesturesSub.running)
                 return;
 
-            gestureSubsystem.Update();
+            gesturesSub.Update();
 
             if (onHoldChanged != null)
             {
-                foreach (var holdGestureEvent in gestureSubsystem.holdGestureEvents)
+                foreach (var holdGestureEvent in gesturesSub.holdGestureEvents)
                     onHoldChanged(holdGestureEvent);
             }
 
             if (onManipulationChanged != null)
             {
-                foreach (var manipulationGestureEvent in gestureSubsystem.manipulationGestureEvents)
+                foreach (var manipulationGestureEvent in gesturesSub.manipulationGestureEvents)
                     onManipulationChanged(manipulationGestureEvent);
             }
 
             if (onNavigationChanged != null)
             {
-                foreach (var navigationGestureEvent in gestureSubsystem.navigationGestureEvents)
+                foreach (var navigationGestureEvent in gesturesSub.navigationGestureEvents)
                     onNavigationChanged(navigationGestureEvent);
             }
 
             if (onTappedChanged != null)
             {
-                foreach (var tappedGestureEvent in gestureSubsystem.tappedGestureEvents)
+                foreach (var tappedGestureEvent in gesturesSub.tappedGestureEvents)
                     onTappedChanged(tappedGestureEvent);
             }
 
             if (onActivate != null)
             {
-                foreach (var activateGestureEvent in gestureSubsystem.activateGestureEvents)
+                foreach (var activateGestureEvent in gesturesSub.activateGestureEvents)
                     onActivate(activateGestureEvent);
             }
         }