+++ Runtime/WindowsMRInternal.cs
@@ -15,7 +15,12 @@ namespace UnityEngine.XR.WindowsMRInternals
     {
         static WindowsMRInternal()
         {
+#if UNITY_EDITOR_WIN
             string pluginFolderPathBase = Path.GetFullPath("Packages/com.unity.xr.windowsmr/Runtime/Plugins/x64");
+#else
+            string pluginFolderPathBase = "";
+#endif
+
             UnityWindowsMR_EmulationLibs_SetPluginFolderPaths(pluginFolderPathBase);
         }
 
@@ -24,8 +29,8 @@ namespace UnityEngine.XR.WindowsMRInternals
         }
 
 #if UNITY_EDITOR_WIN || UNITY_STANDALONE_WIN || UNITY_WINRT
-        [DllImport("WindowsMRXRSDK")]
-        static extern void UnityWindowsMR_EmulationLibs_SetPluginFolderPaths([MarshalAs(UnmanagedType.LPWStr)] string pluginFolderPath_x86_64);
+        [DllImport("WindowsMRXRSDK", CharSet = CharSet.Ansi)]
+        public static extern void UnityWindowsMR_EmulationLibs_SetPluginFolderPaths([MarshalAs(UnmanagedType.LPWStr)] string pluginFolderPath_x86_64);
 #else
         static void UnityWindowsMR_EmulationLibs_SetPluginFolderPaths(string pluginFolderPath_x86_64)
         {